+++ src/App.tsx
@@ -1,27 +1,27 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import { Toaster } from 'react-hot-toast'
 import { AuthProvider } from './contexts/AuthContext'
-import Navbar from './components/Navbar'
-import Home from './pages/Home'
-import OurAnimals from './pages/OurAnimals'
-import JoinUs from './pages/JoinUs'
-import AboutUs from './pages/AboutUs'
-import ContactUs from './pages/ContactUs'
-import Donate from './pages/Donate'
-import LostAndFound from './pages/LostAndFound'
-import AdoptionForm from './pages/OtherForm/AdoptionForm'
-import DonationForm from './pages/OtherForm/DonationForm'
-import VolunteerForm from './pages/OtherForm/VolunteerForm'
-import LostReport from './pages/ReportForm/LostReport'
-import FoundReport from './pages/ReportForm/FoundReport'
-import AbusedReport from './pages/ReportForm/AbusedReport'
-import TransparencyDashboard from './pages/TransparencyDashboard'
-import PasswordReset from './user/components/PasswordReset'
-import ReportModal from './components/ReportModal'
-import LoginModal from './user/components/LoginModal'
-import SignUpModal from './user/components/SignUpModal'
-import ForgotPasswordModal from './user/components/ForgotPasswordModal'
-import Footer from './components/Footer'
+import Navbar from './user/components/Navbar'
+import Home from './user/pages/Main/Home'
+import OurAnimals from './user/pages/Main/OurAnimals'
+import JoinUs from './user/pages/Main/JoinUs'
+import AboutUs from './user/pages/Main/AboutUs'
+import ContactUs from './user/pages/Main/ContactUs'
+import Donate from './user/pages/Main/Donate'
+import LostAndFound from './user/pages/Main/LostAndFound'
+import AdoptionForm from './user/pages/OtherForm/AdoptionForm'
+import DonationForm from './user/pages/OtherForm/DonationForm'
+import VolunteerForm from './user/pages/OtherForm/VolunteerForm'
+import LostReport from './user/pages/ReportForm/LostReport'
+import FoundReport from './user/pages/ReportForm/FoundReport'
+import AbusedReport from './user/pages/ReportForm/AbusedReport'
+import TransparencyDashboard from './user/pages/Main/TransparencyDashboard'
+import PasswordReset from './user/authentication/PasswordReset'
+import ReportModal from './user/components/ReportModal'
+import LoginModal from './user/authentication/LoginModal'
+import SignUpModal from './user/authentication/SignUpModal'
+import ForgotPasswordModal from './user/authentication/ForgotPasswordModal'
+import Footer from './user/components/Footer'
 
 function App() {
   return (
