+++ src/user/pages/Main/LostAndFound.tsx
@@ -2,10 +2,10 @@ import { useState } from 'react'
 import { Search, Filter, MapPin, Calendar, Phone, Mail } from 'lucide-react'
 
 // Import actual images
-import JepoyImage from '../assets/images/Animals/Jepoy.jpg'
-import PutchiImage from '../assets/images/Animals/Putchi.jpg'
-import JoshImage from '../assets/images/Animals/Josh.jpg'
-import MeelooImage from '../assets/images/Animals/Meeloo.jpg'
+import JepoyImage from '../../../assets/images/Animals/Jepoy.jpg'
+import PutchiImage from '../../../assets/images/Animals/Putchi.jpg'
+import JoshImage from '../../../assets/images/Animals/Josh.jpg'
+import MeelooImage from '../../../assets/images/Animals/Meeloo.jpg'
 
 interface LostFoundAnimal {
   id: string
